# spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentAPI Proxy",
    "description": "API for managing agent sessions, schedules, and settings. AgentAPI Proxy provides a unified interface for running AI agents with Kubernetes-based session management.",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/slackbots": {
      "post": {
        "summary": "Create a new SlackBot",
        "description": "Creates a new SlackBot configuration. Each SlackBot corresponds to a Slack App installation. Events are received via Socket Mode (WebSocket).",
        "operationId": "createSlackBot",
        "tags": [
          "SlackBots"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlackBotRequest"
              },
              "examples": {
                "basic": {
                  "summary": "Basic SlackBot",
                  "value": {
                    "name": "My Slack Bot",
                    "session_config": {
                      "initial_message_template": "New Slack message from {{.event.user}} in <#{{.event.channel}}>: {{.event.text}}"
                    }
                  }
                },
                "team-scoped": {
                  "summary": "Team-scoped SlackBot with channel filter",
                  "value": {
                    "name": "Team Bot",
                    "scope": "team",
                    "team_id": "myorg/backend",
                    "bot_token_secret_name": "my-slack-bot-token",
                    "bot_token_secret_key": "bot-token",
                    "allowed_channel_names": [
                      "dev",
                      "backend"
                    ],
                    "session_config": {
                      "initial_message_template": "{{.event.text}}",
                      "tags": {
                        "channel": "{{.event.channel}}"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "SlackBot created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlackBotResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List SlackBots",
        "description": "Returns a list of SlackBots. Non-admin users can see their own SlackBots and team-scoped SlackBots they have access to.",
        "operationId": "listSlackBots",
        "tags": [
          "SlackBots"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by SlackBot status",
            "schema": {
              "$ref": "#/components/schemas/SlackBotStatus"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "description": "Filter by resource scope",
            "schema": {
              "$ref": "#/components/schemas/ResourceScope"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "description": "Filter by team ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of SlackBots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SlackBotResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/slackbots/{id}": {
      "get": {
        "summary": "Get a SlackBot",
        "description": "Returns a specific SlackBot by ID.",
        "operationId": "getSlackBot",
        "tags": [
          "SlackBots"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SlackBot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SlackBot details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlackBotResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - can only access own SlackBots"
          },
          "404": {
            "description": "SlackBot not found"
          }
        }
      },
      "put": {
        "summary": "Update a SlackBot",
        "description": "Updates an existing SlackBot.",
        "operationId": "updateSlackBot",
        "tags": [
          "SlackBots"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SlackBot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSlackBotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SlackBot updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlackBotResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Forbidden - can only update own SlackBots"
          },
          "404": {
            "description": "SlackBot not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a SlackBot",
        "description": "Deletes a SlackBot by ID.",
        "operationId": "deleteSlackBot",
        "tags": [
          "SlackBots"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "SlackBot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "SlackBot deleted successfully"
          },
          "403": {
            "description": "Forbidden - can only delete own SlackBots"
          },
          "404": {
            "description": "SlackBot not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResourceScope": {
        "type": "string",
        "enum": [
          "user",
          "team"
        ],
        "default": "user",
        "description": "Ownership scope. 'user' for personal resources, 'team' for team-shared resources"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "GitHub OAuth token"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
...
}